feat: implement sign-in and sign-up pages with form validation
